feat(continuum): add receipt family witness scope
