+++ docs/architecture/continuum-wesley-role.md
@@ -138,8 +138,10 @@ claim authorship by default.
 
 ## Current Limiting Truth
 
-Wesley now carries an authored `schemas/continuum-receipt-family.graphql`, but
-the current witness-backed minimum surface is still the bounded TTD-plus-Echo
-subset rather than one finished receipt-family contract stack. This note
-defines Wesley's operating role inside Continuum, not the claim that the whole
-Continuum contract surface is already frozen.
+Wesley now carries an authored `schemas/continuum-receipt-family.graphql` and
+a real receipt-family witness scope through
+`wesley witness-continuum --scope receipt-family`. That means the repo can now
+prove one local contract-family stack from authored schema through TTD and Echo
+legs to a bounded conformance witness. This note still does not claim that the
+whole Continuum contract surface is frozen, or that Wesley owns runtime,
+storage, debugger, or substrate semantics outside that bounded proof lane.
