+++ test/fixtures/continuum/receipt-family/boundary.json
@@ -1,0 +1,49 @@
+{
+  "types": {
+    "Capability": [
+      "capabilityId",
+      "name",
+      "scope",
+      "resource",
+      "description"
+    ],
+    "ContinuumReceiptFamilyInvariants": [
+      "_placeholder"
+    ],
+    "DeliveryObservation": [
+      "observationId",
+      "receiptId",
+      "emissionId",
+      "headId",
+      "frameIndex",
+      "sinkId",
+      "outcome",
+      "reason",
+      "observerId",
+      "executionMode",
+      "summary"
+    ],
+    "Receipt": [
+      "receiptId",
+      "headId",
+      "frameIndex",
+      "laneId",
+      "writerId",
+      "inputTick",
+      "outputTick",
+      "admittedRewriteCount",
+      "rejectedRewriteCount",
+      "counterfactualCount",
+      "digest",
+      "summary"
+    ],
+    "Witness": [
+      "witnessId",
+      "receiptId",
+      "kind",
+      "residueHash",
+      "footprint",
+      "summary"
+    ]
+  }
+}
