+++ test/fixtures/continuum/receipt-family/receipt-vs-witness.json
@@ -1,0 +1,30 @@
+{
+  "receiptOnlyFields": [
+    "writerId",
+    "inputTick",
+    "outputTick",
+    "admittedRewriteCount",
+    "rejectedRewriteCount",
+    "counterfactualCount",
+    "digest"
+  ],
+  "witnessOnlyFields": [
+    "witnessId",
+    "kind",
+    "residueHash",
+    "footprint"
+  ],
+  "deliveryObservationForbiddenFields": [
+    "writerId",
+    "inputTick",
+    "outputTick",
+    "admittedRewriteCount",
+    "rejectedRewriteCount",
+    "counterfactualCount",
+    "digest",
+    "witnessId",
+    "kind",
+    "residueHash",
+    "footprint"
+  ]
+}
